+++ database/migrations/2020_01_01_000003_create_plan_subscriptions_table.php
@@ -14,7 +14,7 @@ class CreatePlanSubscriptionsTable extends Migration
      */
     public function up(): void
     {
-        Schema::create(config('cryptodev4.subscriptions.tables.plan_subscriptions'), function (Blueprint $table) {
+        Schema::create(config('laravel-subscriptions.tables.plan_subscriptions'), function (Blueprint $table) {
             $table->increments('id');
             $table->morphs('subscriber');
             $table->integer('plan_id')->unsigned();
@@ -32,7 +32,7 @@ class CreatePlanSubscriptionsTable extends Migration
 
             // Indexes
             $table->unique('slug');
-            $table->foreign('plan_id')->references('id')->on(config('cryptodev4.subscriptions.tables.plans'))
+            $table->foreign('plan_id')->references('id')->on(config('laravel-subscriptions.tables.plans'))
                   ->onDelete('cascade')->onUpdate('cascade');
         });
     }
@@ -44,6 +44,6 @@ class CreatePlanSubscriptionsTable extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists(config('cryptodev4.subscriptions.tables.plan_subscriptions'));
+        Schema::dropIfExists(config('laravel-subscriptions.tables.plan_subscriptions'));
     }
 }