+++ src/Models/Plan.php
@@ -2,21 +2,21 @@
 
 declare(strict_types=1);
 
-namespace CryptoDev4\Subscriptions\Models;
+namespace CryptoDev4\LaravelSubscriptions\Models;
 
-use Spatie\Sluggable\SlugOptions;
-use CryptoDev4\Support\Traits\HasSlug;
-use Spatie\EloquentSortable\Sortable;
+// use Spatie\Sluggable\SlugOptions;
+// use CryptoDev4\Support\Traits\HasSlug;
+// use Spatie\EloquentSortable\Sortable;
 use Illuminate\Database\Eloquent\Model;
-use CryptoDev4\Support\Traits\HasTranslations;
-use CryptoDev4\Support\Traits\ValidatingTrait;
-use Spatie\EloquentSortable\SortableTrait;
+// use CryptoDev4\Support\Traits\HasTranslations;
+// use CryptoDev4\Support\Traits\ValidatingTrait;
+// use Spatie\EloquentSortable\SortableTrait;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 /**
- * CryptoDev4\Subscriptions\Models\Plan.
+ * CryptoDev4\LaravelSubscriptions\Models\Plan.
  *
  * @property int                 $id
  * @property string              $slug
@@ -40,42 +40,41 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
  * @property \Carbon\Carbon|null $created_at
  * @property \Carbon\Carbon|null $updated_at
  * @property \Carbon\Carbon|null $deleted_at
- * @property-read \Illuminate\Database\Eloquent\Collection|\CryptoDev4\Subscriptions\Models\PlanFeature[]      $features
- * @property-read \Illuminate\Database\Eloquent\Collection|\CryptoDev4\Subscriptions\Models\PlanSubscription[] $subscriptions
+ * @property-read \Illuminate\Database\Eloquent\Collection|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature[]      $features
+ * @property-read \Illuminate\Database\Eloquent\Collection|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription[] $subscriptions
  *
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan ordered($direction = 'asc')
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereActiveSubscribersLimit($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereCreatedAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereCurrency($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereDeletedAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereDescription($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereGraceInterval($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereGracePeriod($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereInvoiceInterval($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereInvoicePeriod($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereIsActive($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereName($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan wherePrice($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereProrateDay($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereProrateExtendDue($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereProratePeriod($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereSignupFee($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereSlug($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereSortOrder($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereTrialInterval($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereTrialPeriod($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\Plan whereUpdatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan ordered($direction = 'asc')
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereActiveSubscribersLimit($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereCurrency($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereDeletedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereDescription($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereGraceInterval($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereGracePeriod($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereInvoiceInterval($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereInvoicePeriod($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereIsActive($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereName($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan wherePrice($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereProrateDay($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereProrateExtendDue($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereProratePeriod($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereSignupFee($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereSlug($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereSortOrder($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereTrialInterval($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereTrialPeriod($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\Plan whereUpdatedAt($value)
  * @mixin \Eloquent
  */
-class Plan extends Model implements Sortable
+class Plan extends Model
 {
-    use HasSlug;
     use HasFactory;
     use SoftDeletes;
-    use SortableTrait;
-    use HasTranslations;
-    use ValidatingTrait;
+    // use SortableTrait;
+    // use HasTranslations;
+    // use ValidatingTrait;
 
     /**
      * {@inheritdoc}
@@ -173,9 +172,9 @@ class Plan extends Model implements Sortable
      */
     public function __construct(array $attributes = [])
     {
-        $this->setTable(config('cryptodev4.subscriptions.tables.plans'));
-        $this->mergeRules([
-            'slug' => 'required|alpha_dash|max:150|unique:'.config('cryptodev4.subscriptions.tables.plans').',slug',
+        $this->setTable(config('cryptodev4.laravel-subscriptions.tables.plans'));
+        $this->rules = [
+            'slug' => 'required|alpha_dash|max:150|unique:'.config('cryptodev4.laravel-subscriptions.tables.plans').',slug',
             'name' => 'required|string|strip_tags|max:150',
             'description' => 'nullable|string|max:32768',
             'is_active' => 'sometimes|boolean',
@@ -193,7 +192,7 @@ class Plan extends Model implements Sortable
             'prorate_period' => 'nullable|integer|max:150',
             'prorate_extend_due' => 'nullable|integer|max:150',
             'active_subscribers_limit' => 'nullable|integer|max:100000',
-        ]);
+        ];
 
         parent::__construct($attributes);
     }
@@ -216,13 +215,13 @@ class Plan extends Model implements Sortable
      *
      * @return \Spatie\Sluggable\SlugOptions
      */
-    public function getSlugOptions(): SlugOptions
-    {
-        return SlugOptions::create()
-                          ->doNotGenerateSlugsOnUpdate()
-                          ->generateSlugsFrom('name')
-                          ->saveSlugsTo('slug');
-    }
+    // public function getSlugOptions(): SlugOptions
+    // {
+    //     return SlugOptions::create()
+    //                       ->doNotGenerateSlugsOnUpdate()
+    //                       ->generateSlugsFrom('name')
+    //                       ->saveSlugsTo('slug');
+    // }
 
     /**
      * The plan may have many features.
@@ -231,7 +230,7 @@ class Plan extends Model implements Sortable
      */
     public function features(): HasMany
     {
-        return $this->hasMany(config('cryptodev4.subscriptions.models.plan_feature'), 'plan_id', 'id');
+        return $this->hasMany(config('cryptodev4.laravel-subscriptions.models.plan_feature'), 'plan_id', 'id');
     }
 
     /**
@@ -241,7 +240,7 @@ class Plan extends Model implements Sortable
      */
     public function subscriptions(): HasMany
     {
-        return $this->hasMany(config('cryptodev4.subscriptions.models.plan_subscription'), 'plan_id', 'id');
+        return $this->hasMany(config('cryptodev4.laravel-subscriptions.models.plan_subscription'), 'plan_id', 'id');
     }
 
     /**
@@ -279,7 +278,7 @@ class Plan extends Model implements Sortable
      *
      * @param string $featureSlug
      *
-     * @return \CryptoDev4\Subscriptions\Models\PlanFeature|null
+     * @return \CryptoDev4\LaravelSubscriptions\Models\PlanFeature|null
      */
     public function getFeatureBySlug(string $featureSlug): ?PlanFeature
     {