+++ src/Models/PlanFeature.php
@@ -2,24 +2,24 @@
 
 declare(strict_types=1);
 
-namespace CryptoDev4\Subscriptions\Models;
+namespace CryptoDev4\LaravelSubscriptions\Models;
 
 use Carbon\Carbon;
-use Spatie\Sluggable\SlugOptions;
-use CryptoDev4\Support\Traits\HasSlug;
-use Spatie\EloquentSortable\Sortable;
+// use Spatie\Sluggable\SlugOptions;
+// use CryptoDev4\Support\Traits\HasSlug;
+// use Spatie\EloquentSortable\Sortable;
 use Illuminate\Database\Eloquent\Model;
-use CryptoDev4\Subscriptions\Services\Period;
-use CryptoDev4\Support\Traits\HasTranslations;
-use CryptoDev4\Support\Traits\ValidatingTrait;
-use Spatie\EloquentSortable\SortableTrait;
+use CryptoDev4\LaravelSubscriptions\Services\Period;
+// use CryptoDev4\Support\Traits\HasTranslations;
+// use CryptoDev4\Support\Traits\ValidatingTrait;
+// use Spatie\EloquentSortable\SortableTrait;
 use Illuminate\Database\Eloquent\SoftDeletes;
-use CryptoDev4\Subscriptions\Traits\BelongsToPlan;
+use CryptoDev4\LaravelSubscriptions\Traits\BelongsToPlan;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 /**
- * CryptoDev4\Subscriptions\Models\PlanFeature.
+ * CryptoDev4\LaravelSubscriptions\Models\PlanFeature.
  *
  * @property int                 $id
  * @property int                 $plan_id
@@ -33,34 +33,34 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
  * @property \Carbon\Carbon|null $created_at
  * @property \Carbon\Carbon|null $updated_at
  * @property \Carbon\Carbon|null $deleted_at
- * @property-read \CryptoDev4\Subscriptions\Models\Plan                                                             $plan
- * @property-read \Illuminate\Database\Eloquent\Collection|\CryptoDev4\Subscriptions\Models\PlanSubscriptionUsage[] $usage
+ * @property-read \CryptoDev4\LaravelSubscriptions\Models\Plan                                                             $plan
+ * @property-read \Illuminate\Database\Eloquent\Collection|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscriptionUsage[] $usage
  *
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature byPlanId($planId)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature ordered($direction = 'asc')
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature whereCreatedAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature whereDeletedAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature whereDescription($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature whereId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature whereTitle($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature wherePlanId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature whereResettableInterval($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature whereResettablePeriod($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature whereSlug($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature whereSortOrder($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature whereUpdatedAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanFeature whereValue($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature byPlanId($planId)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature ordered($direction = 'asc')
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature whereDeletedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature whereDescription($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature whereTitle($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature wherePlanId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature whereResettableInterval($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature whereResettablePeriod($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature whereSlug($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature whereSortOrder($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature whereUpdatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanFeature whereValue($value)
  * @mixin \Eloquent
  */
-class PlanFeature extends Model implements Sortable
+class PlanFeature extends Model
 {
-    use HasSlug;
+    // use HasSlug;
     use HasFactory;
     use SoftDeletes;
     use BelongsToPlan;
-    use SortableTrait;
-    use HasTranslations;
-    use ValidatingTrait;
+    // use SortableTrait;
+    // use HasTranslations;
+    // use ValidatingTrait;
 
     /**
      * {@inheritdoc}
@@ -138,17 +138,17 @@ class PlanFeature extends Model implements Sortable
      */
     public function __construct(array $attributes = [])
     {
-        $this->setTable(config('cryptodev4.subscriptions.tables.plan_features'));
-        $this->mergeRules([
-            'plan_id' => 'required|integer|exists:'.config('cryptodev4.subscriptions.tables.plans').',id',
-            'slug' => 'required|alpha_dash|max:150|unique:'.config('cryptodev4.subscriptions.tables.plan_features').',slug',
+        $this->setTable(config('cryptodev4.laravel-subscriptions.tables.plan_features'));
+        $this->rules = [
+            'plan_id' => 'required|integer|exists:'.config('cryptodev4.laravel-subscriptions.tables.plans').',id',
+            'slug' => 'required|alpha_dash|max:150|unique:'.config('cryptodev4.laravel-subscriptions.tables.plan_features').',slug',
             'name' => 'required|string|strip_tags|max:150',
             'description' => 'nullable|string|max:32768',
             'value' => 'required|string',
             'resettable_period' => 'sometimes|integer',
             'resettable_interval' => 'sometimes|in:hour,day,week,month',
             'sort_order' => 'nullable|integer|max:100000',
-        ]);
+        ];
 
         parent::__construct($attributes);
     }
@@ -170,13 +170,13 @@ class PlanFeature extends Model implements Sortable
      *
      * @return \Spatie\Sluggable\SlugOptions
      */
-    public function getSlugOptions(): SlugOptions
-    {
-        return SlugOptions::create()
-                          ->doNotGenerateSlugsOnUpdate()
-                          ->generateSlugsFrom('name')
-                          ->saveSlugsTo('slug');
-    }
+    // public function getSlugOptions(): SlugOptions
+    // {
+    //     return SlugOptions::create()
+    //                       ->doNotGenerateSlugsOnUpdate()
+    //                       ->generateSlugsFrom('name')
+    //                       ->saveSlugsTo('slug');
+    // }
 
     /**
      * The plan feature may have many subscription usage.
@@ -185,7 +185,7 @@ class PlanFeature extends Model implements Sortable
      */
     public function usage(): HasMany
     {
-        return $this->hasMany(config('cryptodev4.subscriptions.models.plan_subscription_usage'), 'feature_id', 'id');
+        return $this->hasMany(config('cryptodev4.laravel-subscriptions.models.plan_subscription_usage'), 'feature_id', 'id');
     }
 
     /**