+++ src/Models/PlanSubscription.php
@@ -2,26 +2,26 @@
 
 declare(strict_types=1);
 
-namespace CryptoDev4\Subscriptions\Models;
+namespace CryptoDev4\LaravelSubscriptions\Models;
 
-use DB;
 use Carbon\Carbon;
 use LogicException;
-use Spatie\Sluggable\SlugOptions;
-use CryptoDev4\Support\Traits\HasSlug;
+// use Spatie\Sluggable\SlugOptions;
+// use CryptoDev4\Support\Traits\HasSlug;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Builder;
-use CryptoDev4\Subscriptions\Services\Period;
-use CryptoDev4\Support\Traits\HasTranslations;
-use CryptoDev4\Support\Traits\ValidatingTrait;
+use CryptoDev4\LaravelSubscriptions\Services\Period;
+// use CryptoDev4\Support\Traits\HasTranslations;
+// use CryptoDev4\Support\Traits\ValidatingTrait;
 use Illuminate\Database\Eloquent\SoftDeletes;
-use CryptoDev4\Subscriptions\Traits\BelongsToPlan;
+use CryptoDev4\LaravelSubscriptions\Traits\BelongsToPlan;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Relations\MorphTo;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Support\Facades\DB;
 
 /**
- * CryptoDev4\Subscriptions\Models\PlanSubscription.
+ * CryptoDev4\LaravelSubscriptions\Models\PlanSubscription.
  *
  * @property int                 $id
  * @property int                 $subscriber_id
@@ -38,41 +38,41 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
  * @property \Carbon\Carbon|null $created_at
  * @property \Carbon\Carbon|null $updated_at
  * @property \Carbon\Carbon|null $deleted_at
- * @property-read \CryptoDev4\Subscriptions\Models\Plan                                                             $plan
- * @property-read \Illuminate\Database\Eloquent\Collection|\CryptoDev4\Subscriptions\Models\PlanSubscriptionUsage[] $usage
+ * @property-read \CryptoDev4\LaravelSubscriptions\Models\Plan                                                             $plan
+ * @property-read \Illuminate\Database\Eloquent\Collection|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscriptionUsage[] $usage
  * @property-read \Illuminate\Database\Eloquent\Model|\Eloquent                                                 $subscriber
  *
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription byPlanId($planId)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription findEndedPeriod()
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription findEndedTrial()
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription findEndingPeriod($dayRange = 3)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription findEndingTrial($dayRange = 3)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription ofSubscriber(\Illuminate\Database\Eloquent\Model $subscriber)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereCanceledAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereCancelsAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereCreatedAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereDeletedAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereDescription($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereEndsAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereTitle($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription wherePlanId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereSlug($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereStartsAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereTrialEndsAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereUpdatedAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereSubscriberId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\Subscriptions\Models\PlanSubscription whereSubscriberType($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription byPlanId($planId)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription findEndedPeriod()
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription findEndedTrial()
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription findEndingPeriod($dayRange = 3)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription findEndingTrial($dayRange = 3)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription ofSubscriber(\Illuminate\Database\Eloquent\Model $subscriber)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereCanceledAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereCancelsAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereDeletedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereDescription($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereEndsAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereTitle($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription wherePlanId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereSlug($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereStartsAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereTrialEndsAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereUpdatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereSubscriberId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\CryptoDev4\LaravelSubscriptions\Models\PlanSubscription whereSubscriberType($value)
  * @mixin \Eloquent
  */
 class PlanSubscription extends Model
 {
-    use HasSlug;
+    // use HasSlug;
     use HasFactory;
     use SoftDeletes;
     use BelongsToPlan;
-    use HasTranslations;
-    use ValidatingTrait;
+    // use HasTranslations;
+    // use ValidatingTrait;
 
     /**
      * {@inheritdoc}
@@ -147,12 +147,12 @@ class PlanSubscription extends Model
      */
     public function __construct(array $attributes = [])
     {
-        $this->setTable(config('cryptodev4.subscriptions.tables.plan_subscriptions'));
-        $this->mergeRules([
+        $this->setTable(config('cryptodev4.laravel-subscriptions.tables.plan_subscriptions'));
+        $this->rules = [
             'name' => 'required|string|strip_tags|max:150',
             'description' => 'nullable|string|max:32768',
-            'slug' => 'required|alpha_dash|max:150|unique:'.config('cryptodev4.subscriptions.tables.plan_subscriptions').',slug',
-            'plan_id' => 'required|integer|exists:'.config('cryptodev4.subscriptions.tables.plans').',id',
+            'slug' => 'required|alpha_dash|max:150|unique:'.config('cryptodev4.laravel-subscriptions.tables.plan_subscriptions').',slug',
+            'plan_id' => 'required|integer|exists:'.config('cryptodev4.laravel-subscriptions.tables.plans').',id',
             'subscriber_id' => 'required|integer',
             'subscriber_type' => 'required|string|strip_tags|max:150',
             'trial_ends_at' => 'nullable|date',
@@ -160,7 +160,7 @@ class PlanSubscription extends Model
             'ends_at' => 'required|date',
             'cancels_at' => 'nullable|date',
             'canceled_at' => 'nullable|date',
-        ]);
+        ];
 
         parent::__construct($attributes);
     }
@@ -188,13 +188,13 @@ class PlanSubscription extends Model
      *
      * @return \Spatie\Sluggable\SlugOptions
      */
-    public function getSlugOptions(): SlugOptions
-    {
-        return SlugOptions::create()
-                          ->doNotGenerateSlugsOnUpdate()
-                          ->generateSlugsFrom('name')
-                          ->saveSlugsTo('slug');
-    }
+    // public function getSlugOptions(): SlugOptions
+    // {
+    //     return SlugOptions::create()
+    //                       ->doNotGenerateSlugsOnUpdate()
+    //                       ->generateSlugsFrom('name')
+    //                       ->saveSlugsTo('slug');
+    // }
 
     /**
      * Get the owning subscriber.
@@ -213,7 +213,7 @@ class PlanSubscription extends Model
      */
     public function usage(): hasMany
     {
-        return $this->hasMany(config('cryptodev4.subscriptions.models.plan_subscription_usage'), 'subscription_id', 'id');
+        return $this->hasMany(config('cryptodev4.laravel-subscriptions.models.plan_subscription_usage'), 'subscription_id', 'id');
     }
 
     /**
@@ -289,7 +289,7 @@ class PlanSubscription extends Model
     /**
      * Change subscription plan.
      *
-     * @param \CryptoDev4\Subscriptions\Models\Plan $plan
+     * @param \CryptoDev4\LaravelSubscriptions\Models\Plan $plan
      *
      * @return $this
      */
@@ -453,7 +453,7 @@ class PlanSubscription extends Model
      * @param string $featureSlug
      * @param int    $uses
      *
-     * @return \CryptoDev4\Subscriptions\Models\PlanSubscriptionUsage
+     * @return \CryptoDev4\LaravelSubscriptions\Models\PlanSubscriptionUsage
      */
     public function recordFeatureUsage(string $featureSlug, int $uses = 1, bool $incremental = true): PlanSubscriptionUsage
     {
@@ -491,7 +491,7 @@ class PlanSubscription extends Model
      * @param string $featureSlug
      * @param int    $uses
      *
-     * @return \CryptoDev4\Subscriptions\Models\PlanSubscriptionUsage|null
+     * @return \CryptoDev4\LaravelSubscriptions\Models\PlanSubscriptionUsage|null
      */
     public function reduceFeatureUsage(string $featureSlug, int $uses = 1): ?PlanSubscriptionUsage
     {